fix sqlite3 transaction handling
Transaction starts, rollbacks, commits were not as intended. By the
current behaviour (as deduced via debugging) the internal value of
autocommit is set to True or at least behaves very similarly to that.
However, access to change said parameter is not exposed until Python
~3.12 or so, despite the underlying sqlite3 version having the option.

Also added a TODO to use autocommit=False once we upgrade to a Python
version that supports it. Would allow removing a lot of boilerplate
code and leaves less ambiguity.